eeg_reader.py combines window_stream.py and process_data.py
