+++ neuros/eeg_reader.py
@@ -1,20 +1,36 @@
 import logging
 from dataclasses import dataclass
+from enum import Enum
 from typing import Iterator, Optional, Tuple
 import numpy as np
 from contextlib import contextmanager
+
+from brainflow import DataFilter, DetrendOperations, FilterTypes
 from brainflow.board_shim import BoardShim, BrainFlowInputParams, BoardIds
 
 logging.basicConfig(level=logging.INFO)
 logger = logging.getLogger(__name__)
 
 
+class Band(Enum):
+    DELTA = (0.5, 4.0)
+    THETA = (4.0, 8.0)
+    ALPHA = (8.0, 13.0)
+    BETA = (13.0, 30.0)
+    GAMMA = (30.0, 100.0)
+    ALL = (0.5, 100.0)
+
+
 @dataclass(frozen=True)
 class WindowConfig:
+    """
+    Configuration for EEG windowing.
+    """
     window_ms: float  # Window size in milliseconds
     overlap_ms: float = 0.0  # Overlap between windows
 
     def __post_init__(self) -> None:
+        """Validate configuration."""
         if self.window_ms <= 0:
             raise ValueError("Window size must be positive")
         if self.overlap_ms >= self.window_ms:
@@ -52,7 +68,16 @@ def create_board_stream(board_id: int = BoardIds.SYNTHETIC_BOARD, params: Option
 
 
 def stream_windows(board: BoardShim, config: WindowConfig) -> Iterator[np.ndarray]:
-    """Generate windows of EEG data from a board stream."""
+    """
+    Stream EEG windows from a board.
+
+    Args:
+        board: BoardShim object.
+        config: WindowConfig object.
+
+    Yields:
+        2D numpy array of EEG data.
+    """
     # Convert ms-based config to samples
     sample_rate = board.get_sampling_rate(board_id=board.board_id)
     window_samples, overlap_samples = config.convert_to_samples(sample_rate)
@@ -92,3 +117,30 @@ def stream_windows(board: BoardShim, config: WindowConfig) -> Iterator[np.ndarra
 
     except GeneratorExit:
         logger.info("Stopping window streaming")
+
+
+def compute_power(channel_data: np.ndarray, sampling_rate: int, band: Band) -> float:
+    """
+    Extract band power from EEG data.
+
+    Args:
+        channel_data: 1D numpy array of samples.
+        sampling_rate: Sampling rate in Hz.
+        band: Brain wave band.
+
+    Returns:
+        Band power as a float.
+    """
+    low_freq, high_freq = band.value
+    filtered = channel_data.copy()
+    DataFilter.detrend(filtered, DetrendOperations.CONSTANT.value)
+    DataFilter.perform_bandpass(
+        filtered,
+        sampling_rate,
+        low_freq,
+        high_freq,
+        4,
+        FilterTypes.BUTTERWORTH.value,
+        0
+    )
+    return np.sqrt(np.mean(np.square(filtered)))
